+++ grain/_src/python/dataset/transformations/packing.py
@@ -45,6 +45,7 @@ class PackingDatasetIterator(dataset.DatasetIterator):
       meta_features: Sequence[str],
       pack_alignment_struct: Any = None,
       padding_struct: Any = None,
+      max_sequences_per_bin: int | None = None,
   ):
     """Initializes the generic packing iterator.
 
@@ -60,6 +61,8 @@ class PackingDatasetIterator(dataset.DatasetIterator):
       meta_features: Meta features that do not require packing.
       pack_alignment_struct: Optional per-feature alignment values.
       padding_struct: Optional per-feature padding values.
+      max_sequences_per_bin: Optional maximum number of input sequences that can
+        be packed into a bin
     """
     super().__init__(parent)
     self._packer_cls = packer_cls
@@ -71,6 +74,7 @@ class PackingDatasetIterator(dataset.DatasetIterator):
     self._meta_features = meta_features
     self._pack_alignment_struct = pack_alignment_struct
     self._padding_struct = padding_struct
+    self._max_sequences_per_bin = max_sequences_per_bin
     self._reset()
 
   def _reset(self):
@@ -159,6 +163,7 @@ class PackingDatasetIterator(dataset.DatasetIterator):
           meta_features=self._meta_features,
           pack_alignment_struct=self._pack_alignment_struct,
           padding_struct=self._padding_struct,
+          max_sequences_per_bin=self._max_sequences_per_bin,
       )
 
   @dataset_stats.record_next_duration_if_output
@@ -216,6 +221,7 @@ class PackingDatasetIterator(dataset.DatasetIterator):
               meta_features=self._meta_features,
               pack_alignment_struct=self._pack_alignment_struct,
               padding_struct=self._padding_struct,
+              max_sequences_per_bin=self._max_sequences_per_bin,
           )
           self._packed_batch_size_bytes = self._current_batch.get_size_bytes()
 
@@ -257,6 +263,7 @@ class PackIterDataset(dataset.IterDataset):
       meta_features: Sequence[str] = (),
       pack_alignment_struct: Any = None,
       padding_struct: Any = None,
+      max_sequences_per_bin: int | None = None,
   ):
     """Initializes the generic packing dataset.
 
@@ -271,6 +278,8 @@ class PackIterDataset(dataset.IterDataset):
       meta_features: Meta features that do not need packing logic.
       pack_alignment_struct: Optional per-feature alignment values.
       padding_struct: Optional per-feature padding values.
+      max_sequences_per_bin: Optional maximum number of input sequences that can
+        be packed into a bin
     """
     super().__init__(parent)
     self._packer_cls = packer_cls
@@ -282,6 +291,13 @@ class PackIterDataset(dataset.IterDataset):
     self._meta_features = meta_features
     self._pack_alignment_struct = pack_alignment_struct
     self._padding_struct = padding_struct
+    self._max_sequences_per_bin = max_sequences_per_bin
+
+    if max_sequences_per_bin is not None and max_sequences_per_bin <= 0:
+      raise ValueError(
+          "If specified, `max_sequences_per_bin` must be a positive integer,"
+          f" but got {max_sequences_per_bin}."
+      )
 
   def __iter__(self) -> dataset.DatasetIterator:
     return PackingDatasetIterator(
@@ -295,6 +311,7 @@ class PackIterDataset(dataset.IterDataset):
         meta_features=self._meta_features,
         pack_alignment_struct=self._pack_alignment_struct,
         padding_struct=self._padding_struct,
+        max_sequences_per_bin=self._max_sequences_per_bin,
     )
 
 
@@ -327,6 +344,7 @@ class FirstFitPackIterDataset(PackIterDataset):
       meta_features: Sequence[str] = (),
       pack_alignment_struct: Any = None,
       padding_struct: Any = None,
+      max_sequences_per_bin: int | None = None,
   ):
     """Creates a dataset that packs sequences using the first-fit strategy.
 
@@ -340,6 +358,8 @@ class FirstFitPackIterDataset(PackIterDataset):
       meta_features: Meta features that do not need packing logic.
       pack_alignment_struct: Optional per-feature alignment values.
       padding_struct: Optional per-feature padding values.
+      max_sequences_per_bin: Optional maximum number of input sequences that can
+        be packed into a bin
     """
     super().__init__(
         parent,
@@ -352,6 +372,7 @@ class FirstFitPackIterDataset(PackIterDataset):
         meta_features=meta_features,
         pack_alignment_struct=pack_alignment_struct,
         padding_struct=padding_struct,
+        max_sequences_per_bin=max_sequences_per_bin,
     )
 
   def __str__(self) -> str:
@@ -380,6 +401,7 @@ class BestFitPackIterDataset(PackIterDataset):
       meta_features: Sequence[str] = (),
       pack_alignment_struct: Any = None,
       padding_struct: Any = None,
+      max_sequences_per_bin: int | None = None,
   ):
     """Creates a dataset that packs sequences using the best-fit strategy.
 
@@ -393,6 +415,8 @@ class BestFitPackIterDataset(PackIterDataset):
       meta_features: Meta features that do not need packing logic.
       pack_alignment_struct: Optional per-feature alignment values.
       padding_struct: Optional per-feature padding values.
+      max_sequences_per_bin: Optional maximum number of input sequences that can
+        be packed into a bin
     """
     super().__init__(
         parent,
@@ -405,6 +429,7 @@ class BestFitPackIterDataset(PackIterDataset):
         meta_features=meta_features,
         pack_alignment_struct=pack_alignment_struct,
         padding_struct=padding_struct,
+        max_sequences_per_bin=max_sequences_per_bin,
     )
 
   def __str__(self) -> str: